dev-env: Setup the minimum basic dev enviroment using Vagrant and Docker
To ease the development workflow for everyone and to make a consistent workflow
vagrant was decided to be used for running docker on Windows. Docker runs
directly on linux and fully supports running containers as well. Windows on the
other hand has a few restrictions, namely Windows Home user cannot run docker.
Also docker uses Windows as its base, which in our case we don't want. NodeJS
and MongoDB run in their own containers isolated from each other. The complete
setup was made using docker-compose as it's much suited for dev enviroments.

There are provisioning and setup scripts made available in tools/ to easily begin
developing the application. Scripts can be welcoming for the new contributors
in the future.

The idea for this was gathered from numerous sources, but mainly from one article
https://hackernoon.com/a-better-way-to-develop-node-js-with-docker-cd29d3a0093