docs: add information regarding dev env for vagrant and linux users
The section for how users need to shutdown and restart the dev env is mentioned.
The vagrant Windows users need to install Virtualbox before they can start the
dev env as the provider.
The ports on Windows vagrant were mapped from 3000:8080. This is now changed and
we have the mapping of 3000:3000 which makes it same across all the platforms.
Linux users need to make the dev env scripts executable as those won't run directly.