rbac: Add role based access control to secured api endpoints.
Fixes #11. I have currently added all the necessary roles, however
haven't added all the permissions and resources that maybe required in the 
future. Resources are self made and specific to our project. Currently the only 
resource that we are working on for now is the `profile` resource and hence only
that has been added. Many more like `post`, `analytics` etc need to be added as 
and when encountered in the future.

All the information needed for the permissions ( updateOwn, readAny ...) can be 
found in the npm module that we are using `accesscontrol`. More info can be 
found out at `https://onury.io/accesscontrol/` on how to use the permissions.

The rbac has been added as an express middleware, and hence can be passed to the
express api route methods and can be used with ease. 
`allowIfLoggedIn` checks if a user is logged in before further processing for 
secured api endpoints.
I have added a middleware for the whole express module which processes each 
request and if has an `authorization` header it passes the loggedInUser as a 
`res.local` var for the next middleware routes. This makes it easier to fetch 
the info about who is the loggedInUser. 
Modified the current api endpoints to work with the newly added rbac
abstractions.

Also added a `vscode` workspace settings so that everyone has the same
codebase and doesn't change file line endings. Added recommendations
for using the same extensions across.