test: Setup testing framework and create test scripts
The testing framework that has been chosen is a combination of
1. `Mocha`- as test runner
2. `Chai` - as preferred BDD assertion library
3. `Sinon` - for stubs and mocks
4. `Instanbul` - for code coverage

To create isolated environments, `tools/run-test` script has been created, 
which also simplifies our task of running tests. We need to pass the OPTION
(unit / integration / coverage). 

The `README.md` has been updated with the steps to run the tests.