userService: Update deleteAllUsers, addUser and minor modifications
`deleteAllUsers` was implemented to physically delete all the users from the
database. Now updated it to set `isDisabled: true` for all the user documents in 
the database. 

`addUser` was accepting `zprn` as a non number and when it threw error Node
returned a Unhandled Exception situation as we were not catching it. Added that
block in the try catch block.